move getGlobeRectangle, restrict clipping to bounding regions
