+++ Cesium3DTiles/src/Tile.cpp
@@ -11,152 +11,11 @@
 #include "TileUtilities.h"
 #include "upsampleGltfForRasterOverlays.h"
 #include <cstddef>
-#include <glm/gtc/matrix_inverse.hpp>
 
 using namespace CesiumAsync;
 using namespace CesiumGeometry;
 using namespace CesiumGeospatial;
 using namespace std::string_literals;
-
-namespace {
-struct GetGlobeRectangle {
-  std::optional<CesiumGeospatial::GlobeRectangle>
-  operator()(const CesiumGeometry::BoundingSphere& boundingSphere) {
-    const glm::dvec3& centerEcef = boundingSphere.getCenter();
-    double radius = boundingSphere.getRadius();
-
-    glm::dmat4 enuToEcef =
-        CesiumGeospatial::Transforms::eastNorthUpToFixedFrame(
-            centerEcef /*, ellipsoid*/);
-    glm::dmat4 ecefBounds = enuToEcef * glm::dmat4(
-                                            radius,
-                                            0.0,
-                                            0.0,
-                                            1.0,
-                                            -radius,
-                                            0.0,
-                                            0.0,
-                                            1.0,
-                                            0.0,
-                                            radius,
-                                            0.0,
-                                            1.0,
-                                            0.0,
-                                            -radius,
-                                            0.0,
-                                            1.0);
-
-    std::optional<CesiumGeospatial::Cartographic> east =
-        CesiumGeospatial::Ellipsoid::WGS84.cartesianToCartographic(
-            ecefBounds[0]);
-    std::optional<CesiumGeospatial::Cartographic> west =
-        CesiumGeospatial::Ellipsoid::WGS84.cartesianToCartographic(
-            ecefBounds[1]);
-    std::optional<CesiumGeospatial::Cartographic> north =
-        CesiumGeospatial::Ellipsoid::WGS84.cartesianToCartographic(
-            ecefBounds[2]);
-    std::optional<CesiumGeospatial::Cartographic> south =
-        CesiumGeospatial::Ellipsoid::WGS84.cartesianToCartographic(
-            ecefBounds[3]);
-
-    if (!east || !west || !north || !south) {
-      return std::nullopt;
-    }
-
-    return CesiumGeospatial::GlobeRectangle(
-        west->longitude,
-        south->latitude,
-        east->longitude,
-        north->latitude);
-  }
-
-  std::optional<CesiumGeospatial::GlobeRectangle>
-  operator()(const CesiumGeometry::OrientedBoundingBox& orientedBoundingBox) {
-    const glm::dvec3& centerEcef = orientedBoundingBox.getCenter();
-    glm::dmat3 enuToEcef =
-        CesiumGeospatial::Transforms::eastNorthUpToFixedFrame(
-            centerEcef /*, ellipsoid*/);
-    const glm::dmat3& halfAxes = orientedBoundingBox.getHalfAxes();
-    glm::dmat3 obbToEnu = glm::affineInverse(enuToEcef) * halfAxes;
-
-    constexpr glm::dvec3 obbVertices[] = {
-        glm::dvec3(1.0, 1.0, 1.0),
-        glm::dvec3(1.0, 1.0, -1.0),
-        glm::dvec3(1.0, -1.0, 1.0),
-        glm::dvec3(1.0, -1.0, -1.0),
-        glm::dvec3(-1.0, 1.0, 1.0),
-        glm::dvec3(-1.0, 1.0, -1.0),
-        glm::dvec3(-1.0, -1.0, 1.0),
-        glm::dvec3(-1.0, -1.0, -1.0)};
-
-    glm::dvec3 vert0Enu = obbToEnu * obbVertices[0];
-
-    int8_t eastmostIndex = 0;
-    double eastmostProjection = vert0Enu.x;
-    int8_t westmostIndex = 0;
-    double westmostProjection = vert0Enu.x;
-    int8_t northmostIndex = 0;
-    double northmostProjection = vert0Enu.y;
-    int8_t southmostIndex = 0;
-    double southmostProjection = vert0Enu.y;
-
-    for (int8_t i = 1; i < 8; ++i) {
-      glm::dvec3 vertEnu = obbToEnu * obbVertices[i];
-      if (vertEnu.x > eastmostProjection) {
-        eastmostProjection = vertEnu.x;
-        eastmostIndex = i;
-      }
-      if (vertEnu.x < westmostProjection) {
-        westmostProjection = vertEnu.x;
-        westmostIndex = i;
-      }
-      if (vertEnu.y > northmostProjection) {
-        northmostProjection = vertEnu.y;
-        northmostIndex = i;
-      }
-      if (vertEnu.y < southmostProjection) {
-        southmostProjection = vertEnu.y;
-        southmostIndex = i;
-      }
-    }
-
-    std::optional<CesiumGeospatial::Cartographic> east =
-        CesiumGeospatial::Ellipsoid::WGS84.cartesianToCartographic(
-            halfAxes * obbVertices[eastmostIndex] + centerEcef);
-    std::optional<CesiumGeospatial::Cartographic> west =
-        CesiumGeospatial::Ellipsoid::WGS84.cartesianToCartographic(
-            halfAxes * obbVertices[westmostIndex] + centerEcef);
-    std::optional<CesiumGeospatial::Cartographic> north =
-        CesiumGeospatial::Ellipsoid::WGS84.cartesianToCartographic(
-            halfAxes * obbVertices[southmostIndex] + centerEcef);
-    std::optional<CesiumGeospatial::Cartographic> south =
-        CesiumGeospatial::Ellipsoid::WGS84.cartesianToCartographic(
-            halfAxes * obbVertices[northmostIndex] + centerEcef);
-
-    if (!east || !west || !north || !south) {
-      return std::nullopt;
-    }
-
-    return CesiumGeospatial::GlobeRectangle(
-        west->longitude,
-        south->latitude,
-        east->longitude,
-        north->latitude);
-  }
-
-  std::optional<CesiumGeospatial::GlobeRectangle>
-  operator()(const CesiumGeospatial::BoundingRegion& boundingRegion) {
-    return boundingRegion.getRectangle();
-  }
-
-  std::optional<CesiumGeospatial::GlobeRectangle>
-  operator()(const CesiumGeospatial::BoundingRegionWithLooseFittingHeights&
-                 boundingRegionWithLooseFittingHeights) {
-    return boundingRegionWithLooseFittingHeights.getBoundingRegion()
-        .getRectangle();
-  }
-};
-} // namespace
 
 namespace Cesium3DTiles {
 
@@ -289,11 +148,10 @@ void Tile::loadContent() {
   // can only be "WEB_MERCATOR" or "GEOGRAPHIC".
   std::set<std::string> projections;
 
-  // const CesiumGeospatial::GlobeRectangle* pRectangle =
-  //    Cesium3DTiles::Impl::obtainGlobeRectangle(&this->getBoundingVolume());
-  // TODO: change name if this works
-  std::optional<CesiumGeospatial::GlobeRectangle> pRectangle =
-      std::visit(GetGlobeRectangle(), this->getBoundingVolume());
+  // TODO: determine the GlobeRectangle for non-region based bounding volumes
+  const CesiumGeospatial::GlobeRectangle* pRectangle =
+      Cesium3DTiles::Impl::obtainGlobeRectangle(&this->getBoundingVolume());
+  bool withinClippingBounds = false;
   if (pRectangle && tileset.supportsRasterOverlays()) {
     // Map overlays to this tile.
     RasterOverlayCollection& overlays = tileset.getOverlays();
@@ -358,7 +216,7 @@ void Tile::loadContent() {
   }
 
   // add geographic texture coordinates for clipping mask
-  bool withinClippingBounds = false;
+  // TODO: generalize this for non-region bounding volumes
   if (pRectangle) {
     const std::vector<std::optional<CesiumGeospatial::GlobeRectangle>>&
         boundingBoxes =
@@ -371,17 +229,7 @@ void Tile::loadContent() {
       }
     }
   }
-/*
-  // convert accumulated projection names to the corresponding projections
-  std::vector<Projection> projections;
-  for (const std::string& projectionName : projections_str) {
-    if (projectionName == "WEB_MERCATOR") {
-      projections.push_back(WebMercatorProjection());
-    } else if (projectionName == "GEOGRAPHIC") {
-      projections.push_back(GeographicProjection());
-    }
-  }
-*/
+
   std::optional<Future<std::shared_ptr<IAssetRequest>>> maybeRequestFuture =
       tileset.requestTileContent(*this);
   if (!maybeRequestFuture) {
@@ -508,52 +356,12 @@ void Tile::loadContent() {
             // TODO factor out into helper function
             // rasterize culling polygons into the tile's bounding rectangle
 
-            //const CesiumGeospatial::GlobeRectangle* pRectangle =
-            //    Cesium3DTiles::Impl::obtainGlobeRectangle(&boundingVolume);
-            const std::optional<CesiumGeospatial::GlobeRectangle> pRectangle =
-                std::visit(GetGlobeRectangle(), boundingVolume);
-            
-            auto val = std::get_if<CesiumGeometry::BoundingSphere>(&boundingVolume);
-            if (val && pRectangle) {
-              
-              SPDLOG_LOGGER_ERROR(
-              pLogger,
-              "East: {}, South: {}, West: {}, North: {}",
-              glm::degrees(pRectangle->getEast()),
-              glm::degrees(pRectangle->getSouth()),
-              glm::degrees(pRectangle->getWest()),
-              glm::degrees(pRectangle->getNorth()));
-              
-              /*
-              for (int i = 0; i < 3; ++i) {
-                SPDLOG_LOGGER_ERROR(
-                pLogger,
-                "HalfAxis[{}]: ({}, {}, {})",
-                i,
-                val->getHalfAxes()[i].x,
-                val->getHalfAxes()[i].y,
-                val->getHalfAxes()[i].z);
-              }*/
-            }
-
+            const CesiumGeospatial::GlobeRectangle* pRectangle =
+                Cesium3DTiles::Impl::obtainGlobeRectangle(&boundingVolume);
 
             if (withinClippingBounds && pRectangle) {
-              // TODO: fix
-              double rectangleWidth =
-                  pRectangle->getEast() - pRectangle->getWest();
-              double rectangleHeight =
-                  pRectangle->getNorth() - pRectangle->getSouth();
-
-              if (rectangleWidth < 0 || rectangleHeight < 0) {
-                SPDLOG_LOGGER_ERROR(
-                pLogger,
-                "East: {}, South: {}, West: {}, North: {}",
-                glm::degrees(pRectangle->getEast()),
-                glm::degrees(pRectangle->getSouth()),
-                glm::degrees(pRectangle->getWest()),
-                glm::degrees(pRectangle->getNorth()));
-              };
-              
+              double rectangleWidth = pRectangle->computeWidth();
+              double rectangleHeight = pRectangle->computeHeight();
 
               // Create opacity mask texture
 
@@ -1107,12 +915,8 @@ Tile::generateTextureCoordinates(
 
   // Generate texture coordinates for each projection.
   if (!projections.empty()) {
-    /*
     const CesiumGeospatial::GlobeRectangle* pRectangle =
         Cesium3DTiles::Impl::obtainGlobeRectangle(&boundingVolume);
-    */
-    std::optional<CesiumGeospatial::GlobeRectangle> pRectangle =
-        std::visit(GetGlobeRectangle(), boundingVolume);
     if (pRectangle) {
       for (const std::string& projectionName : projections) {
         Projection projection;
